1: Created two files in ~/.config/rofi 	- Added a config and theme file for rofi's power menu. 	- Purpose is to not use the "entry" input box 	  and try to streamline the menus look. 2: Edited Waybar powermenu script file 	- Changed rofi run command to use theme_powermenu 3: Changes to hyprland.conf 	- Added $powerMenu and a bind using it so that 	  you can call powermenu.sh without interacting with waybar. 	  This can be seen as a similar feature to Super+L 	  on Windows to lock the pc. 	-Added bind for Alt+Tab to switch between workspaces 	 like in Windows. This will only switch to workspaces 	 with a program in it and it does loop back around. 	-Added firefox to My Programs section as $webBrowser 	-Changed keybinding for thunar to use its existing $fileManager 	 variable which was unused(?)
I based my hyprland setup on dionysus however due to some waybar settings
and not having an ASUS laptop I had to do a lot of changes for my own setup
and therefore I have backported the changes above to dionysus. In short, I
do not know for 100% that my change of removing 'entry' in rofi powermenu
will not cause an unwanted appearance. I left code for entry intact to be
easily enabled again if the look is unwanted.